Google Maps Api Data Save feature part 2
Google maps data sent in an array using a JSON Content-type header will be taken in, the data checked against the database, then saved if not already existing. Process could be faster, but is async, so other things can be processed at the same time.

Next step is to get the check request as one request for all the data in the array, rather than one request per object in the array, to save time and lower the amount of requests on the database.